requirements.txt
alembic==1.7.6
anyio==3.4.0
argcomplete==1.12.3
asgiref==3.4.1
attrs==21.4.0
bcrypt==3.2.0
Beaker==1.10.0
beautifulsoup4==4.9.3
blivet==3.4.2
blivet-gui==2.3.0
Brlapi==0.8.2
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.4
chrome-gnome-shell==0.0.0
click==8.0.3
cryptography==36.0.1
cupshelpers==1.0
dasbus==1.6
DateTime==4.3
dbus-python==1.2.18
distro==1.6.0
Django==4.0.2
django-filter==21.1
djangorestframework==3.13.1
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.70.0
fedora-third-party==0.8
fros==1.1
gpg==1.15.1
greenlet==1.1.2
h11==0.12.0
humanize==0.5.1
idna==3.3
iniconfig==1.1.1
Jinja2==3.0.3
langtable==0.0.56
libcomps==0.1.18
lxml==4.6.5
Mako==1.1.4
Markdown==3.3.6
MarkupSafe==2.0.1
mod-wsgi==4.9.0
nftables==0.1
olefile==0.46
packaging==21.3
passlib==1.7.4
Paste==3.5.0
pexpect==4.8.0
pid==2.2.3
Pillow==8.3.2
pluggy==1.0.0
ply==3.11
productmd==1.33
psycopg2-binary==2.9.2
ptyprocess==0.6.0
pwquality==1.4.4
py==1.11.0
pyasn1==0.4.8
pycairo==1.20.1
pycparser==2.21
pycrypto==2.6.1
pycups==2.0.1
pycurl==7.44.1
pydantic==1.8.2
pyenchant==3.2.2
PyGObject==3.42.0
pykickstart==3.34
pyOpenSSL==21.0.0
pyparsing==3.0.7
pyparted==3.11.7
PySocks==1.7.1
pytest==6.2.5
python-augeas==1.1.0
python-dateutil==2.8.1
python-dotenv==0.19.2
python-jose==3.3.0
python-meh==0.50
python-multipart==0.0.5
pytz==2021.3
pyudev==0.22.0
pyxdg==0.27
regex==2022.1.18
requests==2.27.0
requests-file==1.5.1
requests-ftp==0.3.1
rpm==4.17.0
rsa==4.8
selinux==3.3
sepolicy==3.3
setools==4.4.0
simpleaudio==1.0.4
simpleline==1.8.2
six==1.16.0
sniffio==1.2.0
sockets==1.0.0
sos==4.2
soupsieve==2.3.1
SQLAlchemy==1.4.27
sqlparse==0.4.2
starlette==0.16.0
systemd-python==234
Tempita==0.5.2
toml==0.10.2
typing==3.7.4.3
typing_extensions==4.0.0
urllib3==1.26.7
uvicorn==0.15.0
zope.interface==5.4.0
